+++ packages/components/src/components/carousel/carousel.test.ts
@@ -490,7 +490,6 @@ describe('<sd-carousel>', () => {
 
         // Assert
         expect(el.scrollContainer).to.have.attribute('aria-busy', 'false');
-        expect(el.scrollContainer).to.have.attribute('aria-atomic', 'true');
 
         expect(pagination).to.have.attribute('aria-controls', el.scrollContainer.id);
         for (const paginationItem of pagination.querySelectorAll('.carousel__pagination-item')) {
@@ -525,7 +524,6 @@ describe('<sd-carousel>', () => {
 
         // Assert
         expect(el.scrollContainer).to.have.attribute('aria-busy', 'false');
-        expect(el.scrollContainer).to.have.attribute('aria-atomic', 'true');
 
         expect(pagination).to.have.attribute('role', 'tablist');
         expect(pagination).to.have.attribute('aria-controls', el.scrollContainer.id);
@@ -796,7 +794,7 @@ describe('<sd-carousel>', () => {
     it('should resume if the user clicks the resume button', async () => {
       // Arrange
       const el = await fixture<SdCarousel>(html`
-        <sd-carousel autoplay>
+        <sd-carousel autoplay loop>
           <sd-carousel-item>Node 1</sd-carousel-item>
           <sd-carousel-item>Node 2</sd-carousel-item>
           <sd-carousel-item>Node 3</sd-carousel-item>
@@ -848,6 +846,28 @@ describe('<sd-carousel>', () => {
 
       // Assert
       expect(el.goToSlide).to.have.been.calledWith(slidesPerMove);
+    });
+
+    it('should unblock autoplay when a dot is clicked', async () => {
+      // Arrange
+      const el = await fixture<SdCarousel>(html`
+        <sd-carousel autoplay variant="dot">
+          <sd-carousel-item>Node 1</sd-carousel-item>
+          <sd-carousel-item>Node 2</sd-carousel-item>
+          <sd-carousel-item>Node 3</sd-carousel-item>
+        </sd-carousel>
+      `);
+      sinon.stub(el, 'next');
+
+      await el.updateComplete;
+
+      // Act
+      const secondDot = el.shadowRoot!.querySelectorAll('.carousel__pagination-item')[1] as HTMLElement;
+      secondDot.click();
+      await el.updateComplete;
+
+      // Assert
+      expect(secondDot).to.not.have.focus;
     });
   });
 });