+++ packages/components/src/components/textarea/textarea.ts
@@ -403,6 +403,7 @@ export default class SdTextarea extends SolidElement implements SolidFormControl
               enterkeyhint=${ifDefined(this.enterkeyhint)}
               inputmode=${ifDefined(this.inputmode)}
               aria-describedby="help-text invalid-message"
+              aria-invalid=${this.showInvalidStyle}
               @change=${this.handleChange}
               @input=${this.handleInput}
               @invalid=${this.handleInvalid}