+++ packages/components/src/utilities/localize.ts
@@ -74,6 +74,8 @@ export interface Translation extends DefaultTranslation {
   $name: string; // e.g. English, Español
   $dir: 'ltr' | 'rtl';
 
+  autoplay: string;
+  audioPlayer: string;
   carousel: string;
   clearEntry: string;
   close: string;
@@ -82,8 +84,12 @@ export interface Translation extends DefaultTranslation {
   goToSlide: (slide: number, count: number) => string;
   hidePassword: string;
   loading: string;
+  mute: string;
   nextSlide: string;
   numOptionsSelected: (num: number) => string;
+  pauseAudio: string;
+  playAudio: string;
+  playbackSpeed: string;
   previousSlide: string;
   progress: string;
   remove: string;
@@ -92,15 +98,10 @@ export interface Translation extends DefaultTranslation {
   scrollToStart: string;
   selectAColorFromTheScreen: string;
   selectDefaultPlaceholder: string;
-  showPassword: string;
-  slideNum: (slide: number) => string;
-  toggleColorFormat: string;
-  audioPlayer: string;
-  playAudio: string;
-  pauseAudio: string;
-  mute: string;
-  unmute: string;
-  showMore: string;
   showLess: string;
-  playbackSpeed: string;
+  showMore: string;
+  showPassword: string;
+  slideNum: (slide: number, count: number) => string;
+  toggleColorFormat: string;
+  unmute: string;
 }