+++ packages/placeholders/CHANGELOG.md
@@ -1,3 +1,23 @@
+# [@solid-design-system/placeholders-v2.0.1](https://github.com/solid-design-system/solid/compare/placeholders/2.0.0...placeholders/2.0.1) (2024-11-27)
+
+
+### Bug Fixes
+
+* improve sd-video a11y ([#1644](https://github.com/solid-design-system/solid/issues/1644)) ([6dec218](https://github.com/solid-design-system/solid/commit/6dec2189231b9d6538391b8e37b5228d4253d9ed)), closes [#1494](https://github.com/solid-design-system/solid/issues/1494) [#1529](https://github.com/solid-design-system/solid/issues/1529)
+
+# [@solid-design-system/placeholders-v2.0.0](https://github.com/solid-design-system/solid/compare/placeholders/1.8.0...placeholders/2.0.0) (2024-11-18)
+
+
+* chore!: replace ui-placeholder-video with sds-placeholder-video (#1657) ([77f8343](https://github.com/solid-design-system/solid/commit/77f834338493d3d1e2a79bd23519477629a4edaa)), closes [#1657](https://github.com/solid-design-system/solid/issues/1657) [#1644](https://github.com/solid-design-system/solid/issues/1644)
+
+
+### BREAKING CHANGES
+
+* The `ui-placeholder-video.mp4` has been deleted and
+replaced by `sds-placeholder-video.webm`.
+
+This PR addresses replacing the current placeholder video with the SDS
+
 # [@solid-design-system/placeholders-v1.8.0](https://github.com/solid-design-system/solid/compare/placeholders/1.7.1...placeholders/1.8.0) (2024-11-08)